validation for Longitude and Latitude in PlaceQueryDTO

+++ src/main/java/pl/warkoczewski/foodspot/dto/DisplayPlaceDTO.java
@@ -3,11 +3,13 @@ package pl.warkoczewski.foodspot.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @Data
 public class DisplayPlaceDTO {
-    private Double lon;
-    private Double lat;
+    private BigDecimal lon;
+    private BigDecimal lat;
     private String name;
     private String address;
